OCS-444 - intermittent commit, storing exponents of converter functions

+++ src/dynamic.hpp
@@ -60,7 +60,7 @@ namespace munits {
 
 
     const std::vector<munits::Metric> & GetMatrix();
-    const int GetMatrixIndex(std::vector<int> & searched);
+    const long long int GetMatrixIndex(std::vector<int> &searched);
     const std::map<std::string, const std::shared_ptr<ConverterFunction>> & GetPrefixes();
